requirements.txt
fastapi==0.115.0
uvicorn==0.31.0
python-docx==1.1.2
docx2pdf==0.1.8
reportlab==4.2.2
openpyxl==3.1.5
pandas==2.2.3
Pillow==11.0.0
python-multipart==0.0.9
pyzipper==0.3.6
gunicorn==23.0.0
pypandoc-binary==1.13
mammoth==1.6.0
xhtml2pdf==0.2.16
pdf2image==1.17.0